+++ lib/src/auth/AuthInterface.dart
@@ -1,0 +1,13 @@
+part of route_provider;
+
+class AuthResponse {}
+
+abstract class Auth {
+    Future<AuthResponse> isAuthed(HttpRequest request, Map params);
+}
+
+class StaticAuth implements Auth {
+    bool authed;
+    StaticAuth({this.authed});
+    Future<AuthResponse> isAuthed(HttpRequest request, Map params) async => this.authed ? new AuthResponse() : null;
+}