+++ CHANGELOG.md
@@ -1,4 +1,16 @@
-## 0.2.0 (2015-08-??)
+## 0.3.3 (2015-10-16)
+- change auth signature and handling
+
+## 0.3.2 (2015-10-16)
+- update tests
+
+## 0.3.1 (2015-10-16)
+- parsing http-request and params to auth handler
+
+## 0.3.0 (2015-10-16)
+- add the optional named parameter 'auth' (interface Auth) to the route method for authentication possibilities. it checks before the controller and responsehandler do there work if the call is authenticated - default is true.
+
+## 0.2.0 (2015-08-27)
 - rename RouteControllerEmpty to EmptyRouteController
 - add JsonResponser
 - add ApiRestController